requirements.txt
pytest==7.4.3
selenium==4.15.2
flaky
allure-pytest